+++ src/Room/index.tsx
@@ -2,10 +2,11 @@ import * as React from 'react';
 import { connect } from 'react-redux'
 import system from '@rebass/components'
 
-import { actions } from '../models/user'
+import { getSingleton, QUEUES_CHANNEL } from '../cable'
+import { actions as userActions } from '../models/user'
 
 import { State as RootState } from '../reducers'
-import { State, types } from './redux'
+import { actions, State, types } from './redux'
 
 import ReactPlayer from 'react-player'
 
@@ -15,9 +16,12 @@ const Container = system({
   flex: '1',
 })
 
-type Props = State & typeof actions
+type Props = State & typeof userActions
 class Room extends React.Component<Props, {}> {
   componentWillMount() {
+    // Actions sent to subscribe are dispatched by the cable saga
+    // and should not be previously wrapped in dispatch
+    getSingleton().subscribe(QUEUES_CHANNEL, 'room', actions.receiveSongs)
     this.props.getUsers(types.GET_USERS_OK, types.GET_USERS_ERR)
   }
 
@@ -40,7 +44,7 @@ class Room extends React.Component<Props, {}> {
 type MapStateToProps = (state: RootState) => State
 const mapStateToProps: MapStateToProps = (state) => state.room
 
-export default connect<State, typeof actions, {}>(
+export default connect<State, typeof userActions, {}>(
   mapStateToProps,
-  actions,
+  userActions,
 )(Room)
