+++ src/Room/redux/index.ts
@@ -1,2 +1,3 @@
+export { default as actions } from './actions'
 export { default as reducer } from './reducers'
 export { State, types } from './types'
