+++ src/Room/redux/reducers.ts
@@ -1,5 +1,4 @@
 import { Action, types, State } from './types'
-import { types as songTypes } from '../../models/song'
 
 export const initialState: State = {
   users: [],
@@ -10,7 +9,7 @@ export default function reducer(state: State = initialState, action: Action) {
   switch (action.type) {
     case(types.GET_USERS_OK):
       return { users: action.users, songs: state.songs }
-    case(songTypes.RECEIVE_SONGS):
+    case(types.RECEIVE_SONGS):
       return { users: state.users, songs: action.songs }
     default:
       return state
