+++ src/models/song/index.ts
@@ -1,3 +1,2 @@
-export { default as actions } from './actions'
 export * from './deserializers'
-export { ActionCreators, Song, types, Action } from './types'
+export { Song } from './types'
